Removed empty directory in initial setup
The creation of the empty manuscript/resources/images directory, with
symlinks to it from other places within the repo, was breaking Leanpub's
book-generation process since empty directories are not stored in Git,
so there were broken symlinks in the repo.

Now ox-leanpub-book creates a placeholder README file inside
manuscript/resources/images, to ensure the directory is tracked by Git.
The file also contains an explanation and instructions for how to
prevent the problem if it needs to be removed.