Improvements and docs for special block handling
Org-to-Markua block mappings are now defined in the new
`org-leanpub-markua-block-mapping` variable instead of a defun-local
variable.

New option `#+OX_MARKUA_EXPORT_TYPE` which can be set to "book" (default
value) or "course" to indicate whether the current file will be exported
as a book or a course. At this moment the only difference is how
EXERCISE blocks are handled, see below.

Blocks of type EXERCISE are exported either as `X>` (with rendered
contents) or as `{exercise}` (with as-is content), depending on the
value of OX_MARKUA_EXPORT_TYPE.

Improved documentation.